+++ lib/screens/place/controls/interior.dart
